cppcheck Bundle 57: %d/%u + invalidLifetime + uninitvar + nullPointer sweep
Eleven cppcheck warning/error sites across seven Linux-desktop source
files, all unrelated, batched as the post-Bundle-56 narrow-scope sweep.

invalidPrintfArgType_sint (11 sites, 4 files): `%d` -> `%u` for unsigned
args on the format strings used to build display/shader-preset/save-state
identifiers. Wire format unchanged in practice (values never reach
INT_MAX), the format now matches the actual arg type:
  - gfx/display_servers/dispserv_kms.c:104,109 (x6) — resolution_list_qsort_func
    builds `"%04ux%04u (%u Hz)"` over `video_display_config_t::{width,height,
    refreshrate}` (all `unsigned` per video_display_server.h:36-39).
  - gfx/video_shader_parse.c:255,263,271 (x3) — VID-USER-ROT / VID-FINAL-ROT /
    SCREEN-ORIENT preset wildcards format `unsigned * 90` derived from
    settings->uints.video_rotation, retroarch_get_rotation() (unsigned int per
    retroarch.h:196), or settings->uints.screen_orientation.
  - gfx/video_crt_switch.c:218 — `super_width` is `unsigned int` per
    crt_sr2_init signature at line 164.
  - command.c:1841 — `del_idx` is `unsigned` (line 1674: `unsigned del_idx
    = UINT_MAX`). Cast to `(int)` rather than swap to `%u` because the
    bounded buffer (snprintf size=5) made -Wformat-truncation= fire on `%u`
    due to GCC's tighter unsigned range analysis; `(int)` cast keeps the
    format-arg type match cppcheck wanted while preserving the prior
    int-range bound. Behaviour identical for the realistic savestate-slot
    range (del_idx <= savestate_max_keep, single/double-digit in practice).

invalidLifetime (1 site, 1 file): `gfx/gfx_thumbnail.c:1177` —
`coords.color = (const float*)thumbnail_color` was reassigning over a
prior `coords.color = (const float*)shadow_color` set inside the inner
shadow-drawing block (line 1163); `shadow_color[16]` was declared at the
top of that inner block (was line 1129) and goes out of scope at the
block-end three lines before the outer reassign. Strictly UB in C99
6.2.4 even without a read of the dangling pointer. Fix: hoisted the
`shadow_color[16]` declaration to the outer scope alongside
`thumbnail_color[16]` (around line 1016), so the array stays in scope
across the reassign. The brace-initializer `{0,0,0,1, ...}` is identical
in both positions; `gfx_display_set_alpha(shadow_color, shadow_alpha)`
in the inner block still mutates only the alpha channels as before. The
shadow branch isn't exercised when `shadow == NULL` so the now-outer
array decl is dead bytes (64 bytes on stack) — acceptable; the function
runs once per thumbnail draw, not in a hot inner loop.

uninitvar (1 site, 1 file): `gfx/drivers_context/xegl_ctx.c:299` —
`temp.visualid = vid` reads `EGLint vid` which is only initialized by
`egl_get_native_visual_id(&xegl->egl, &vid)` under `#ifdef HAVE_EGL`
(line 294-297). On Linux desktop XEGL is always built with HAVE_EGL but
the structure leaves the read unsafe-by-default. Fix: initialize `vid =
0` at declaration (line 271, split from the comma-decl with `num_visuals`
to keep the assignment readable). Cost: one branch evaluation; gain: well-
defined behavior if HAVE_EGL is ever unset for this driver.

nullPointerRedundantCheck (1 site, 1 file): `input/input_driver.c:4523`
— `input_config_get_bind_string` derefs `bind->mbutton` without a NULL
check, but the four feeder branches at lines 4478/4482/4486/4490 all
guard `bind &&`/`auto_bind &&`, and the `#ifndef RARCH_CONSOLE` block at
4499 explicitly NULL-checks `bind`. So callers may legitimately pass
`bind = NULL` with `auto_bind != NULL` and reach line 4523 with a NULL
deref. Fix: `if (bind && bind->mbutton != NO_BTN)`. Same pattern as the
upstream callers' guards; no logic change for the (overwhelmingly common)
non-NULL `bind` case.

False positives confirmed and not fixed:
  - input/input_driver.c:2134 arrayIndexOutOfBoundsCond — cppcheck warns
    `idx<2` either redundant or array `input_remap_ids[16][37]` accessed
    at index 92. The `if (idx < 2 && id < 2)` IS doing real bounds work:
    max offset = RARCH_FIRST_CUSTOM_BIND(16) + (1*4) + (1*2) = 22, well
    under RARCH_CUSTOM_BIND_LIST_END(37); index 92 is cppcheck's worst-
    case for the unbounded case but the bound prevents it.
  - gfx/drivers_shader/shader_gl_cg.c:917 arrayIndexOutOfBoundsCond —
    Cg shaders are HAVE_CG-gated (deprecated by NVIDIA); deferred.

Build: 7 buildable files all compile clean under `make -j4` with
`-Wall -Wsign-compare`. cppcheck cluster cleared on all 14 fixed sites
post-fix. Cluster too narrow (5 unrelated rule families across 7 files)
for an AUDIT-POLICY entry.

+++ command.c
@@ -1838,7 +1838,7 @@ static void command_scan_states(
       strlcpy(s, savefile_root, savefile_root_length + 1);
       /* ".state0" is just ".state" instead, so don't print that. */
       if (del_idx > 0)
-         snprintf(s + savefile_root_length, 5, "%d", del_idx);
+         snprintf(s + savefile_root_length, 5, "%d", (int)del_idx);
    }
 
    dir_list_free(dir_list);
